consolidate apptopBar into sidebar pure login
