+++ web/scenes/PortalV3/layout/Shell/NavItem.tsx
@@ -2,24 +2,33 @@ import { Link } from "@/components/Link";
 import { ReactNode } from "react";
 import { twMerge } from "tailwind-merge";
 
+export const sidebarNavItemClassName = (props?: {
+  active?: boolean;
+  dimmed?: boolean;
+  className?: string;
+}) =>
+  twMerge(
+    "flex h-10 min-w-0 items-center gap-2 rounded-[10px] pl-3 pr-4 font-world text-13 font-normal leading-none outline-none transition-colors",
+    "focus-visible:ring-2 focus-visible:ring-grey-300 focus-visible:ring-offset-2 focus-visible:ring-offset-portal-canvas",
+    props?.active
+      ? "border border-portal-border bg-white text-portal-text shadow-portal-card"
+      : "text-portal-muted hover:bg-portal-border hover:text-portal-text",
+    props?.dimmed && "opacity-40",
+    props?.className,
+  );
+
 export const NavItem = (props: {
   href: string;
   label: string;
   icon?: ReactNode;
   trailing?: ReactNode;
   active?: boolean;
+  current?: boolean;
   dimmed?: boolean;
+  className?: string;
 }) => {
-  const { href, label, icon, trailing, active, dimmed } = props;
-
-  const className = twMerge(
-    "flex h-10 min-w-0 items-center gap-2 rounded-[10px] pl-3 pr-4 font-world text-13 font-normal leading-none outline-none transition-colors",
-    "focus-visible:ring-2 focus-visible:ring-grey-300 focus-visible:ring-offset-2 focus-visible:ring-offset-portal-canvas",
-    active
-      ? "border border-portal-border bg-white text-portal-text shadow-portal-card"
-      : "text-portal-muted hover:bg-portal-border hover:text-portal-text",
-    dimmed && "opacity-40",
-  );
+  const { href, label, icon, trailing, active, current, dimmed, className } =
+    props;
 
   const content = (
     <>
@@ -38,8 +47,8 @@ export const NavItem = (props: {
   return (
     <Link
       href={href}
-      className={className}
-      aria-current={active ? "page" : undefined}
+      className={sidebarNavItemClassName({ active, dimmed, className })}
+      aria-current={current === false ? undefined : active ? "page" : undefined}
     >
       {content}
     </Link>
